merge

+++ src/main/java/org/nustaq/heapoff/OffHeapByteTree.java
@@ -13,7 +13,7 @@ import org.nustaq.heapoff.bytez.malloc.MallocBytezAllocator;
 public class OffHeapByteTree {
 
     static int estimateMBytesForIndex(int keylen, int numOfKeys) {
-        return (int) Math.max(10, (2 * keylen * numOfKeys) / 1000 / 1000 );
+        return Math.max(10, (2 * keylen * numOfKeys) / 1000 / 1000 );
     }
 
     MallocBytezAllocator alloc = new MallocBytezAllocator();
